requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
black==24.8.0
    # via -r requirements.in
certifi==2024.8.30
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via black
idna==3.10
    # via requests
mypy-extensions==1.0.0
    # via black
packaging==24.1
    # via black
pathspec==0.12.1
    # via black
platformdirs==4.3.6
    # via black
python-dotenv==1.0.1
    # via -r requirements.in
redis==5.0.8
    # via spotipy
requests==2.32.3
    # via spotipy
slack-sdk==3.33.1
    # via -r requirements.in
spotipy==2.24.0
    # via -r requirements.in
urllib3==2.2.3
    # via
    #   requests
    #   spotipy
